Combining audio files feature and alerts for successful combination and save added. Audiofile name conflict also generates an alert.

+++ Assignment3/src/application/CreateMenuController.java
@@ -1,5 +1,6 @@
 package application;
 
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -12,6 +13,8 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.MenuButton;
@@ -50,7 +53,6 @@ public class CreateMenuController {
 	
 	private String _selectedText;
 	private String _voiceSelection;
-
 	private Label labelValue;
 	
 
@@ -112,17 +114,33 @@ public class CreateMenuController {
 	
 	@FXML
 	private void handleButtonSave() throws IOException {
+		String audiofileName = textFieldAudioName.getText();
+		//checking if audiofile with same name already exists
+		File tempFile = new File("audiofiles/"+audiofileName+".wav");
+		if(tempFile.exists()) {
+			Alert alert = new Alert(AlertType.WARNING);
+			alert.setTitle("Already exists");
+			alert.setHeaderText(null);
+			alert.setContentText("Audio with same name already exists, please choose a new name");
+			alert.showAndWait();
+			return;
+		}
 		FileWriter scmwriter = new FileWriter("temporaryfiles/audiofile.scm", false);
 		scmwriter.write(_voiceSelection);
 		scmwriter.close();
 		_selectedText = textAreaResults.getSelectedText();
-		String audiofileName = textFieldAudioName.getText();
 		FileWriter writer = new FileWriter("temporaryfiles/audiotext", false);
 		writer.write(_selectedText);
 		writer.close();
 		ProcessBuilder pb1 = new ProcessBuilder();
-		pb1.command("bash", "-c", "text2wave temporaryfiles/audiotext -o audiofiles/" + audiofileName + ".wav -eval audiofile.scm");
-		Process process1 = pb1.start();
+		pb1.command("bash", "-c", "text2wave temporaryfiles/audiotext -o audiofiles/" + audiofileName + ".wav -eval temporaryfiles/audiofile.scm");
+		Process process = pb1.start();
+		Alert alert = new Alert(AlertType.INFORMATION);
+		alert.setTitle("Successfully created");
+		alert.setHeaderText(null);
+		alert.setContentText("Your audio file has been saved");
+		alert.showAndWait();
+		return;
 	}
 	
 	@FXML
